requirements.txt
# since it's a --pre needs the gte to work
azure-monitor-opentelemetry>=1.0.0b1
Django==4.2.3
python-dotenv==1.0.0
graphene-django==3.1.3

pytest==7.4.0
pytest-django==4.5.2
pytest-asyncio==0.21.1
pylint==2.17.4
coverage==7.2.7
black==23.7.0
pytest-cov==4.1.0
